refactor : create Default Response & Controller Advice

+++ backend/src/main/java/com/a105/api/service/FriendshipService.java
@@ -1,6 +1,6 @@
 package com.a105.api.service;
 
-import com.a105.api.request.FriendRequest;
+import com.a105.api.request.FriendshipRequest;
 import com.a105.api.response.FriendInfoResponse;
 import com.a105.domain.friendship.FriendshipDto;
 import com.a105.domain.friendship.Friendship;
@@ -12,8 +12,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,7 +22,7 @@ public class FriendshipService {
     private final FriendshipRepository friendRepository;
     private final UserService userService;
     public List<FriendInfoResponse> getFriendDtos(Long id){
-        List<FriendshipDto> friendshipDtos = friendRepository.findFriendDtos(id);
+        List<FriendshipDto> friendshipDtos = friendRepository.findFriendshipDtos(id);
         List<FriendInfoResponse> friendInfoResponses = new ArrayList<>();
 
         for (FriendshipDto friendshipDto:friendshipDtos) {
@@ -56,6 +54,25 @@ public class FriendshipService {
         }
     }
 
+    /**
+     * 친구를 삭제한다.
+     * @param friendId
+     */
+    public void deleteFriend(Long userId, Long friendId){
+        Optional<Friendship> sent = findByUserIdAndFriendId(userId, friendId);
+        Optional<Friendship> received = findByUserIdAndFriendId(friendId, userId);
+
+        if(sent.isPresent() && sent.get().getStatus() == 1){
+            friendRepository.delete(sent.get());
+            return;
+        } else if (received.isPresent() && received.get().getStatus() == 1){
+            friendRepository.delete(received.get());
+            return;
+        } else {
+            throw new BadRequestException("친구를 삭제할 수 없습니다.");
+        }
+    }
+
     public Friendship findById(Long id) {
         return friendRepository.findById(id)
             .orElseThrow(() -> new ResourceNotFoundException("Friendship", "id", id));
@@ -65,29 +82,48 @@ public class FriendshipService {
         return friendRepository.findByUserIdAndFriendId(userId, friendId);
     }
 
-
-    public void declineReceivedRequest(Long id, FriendRequest friendRequest){
-        // 받은 요청을 거절한다.
+    /**
+     * 받은 요청을 거절한다.
+     * @param friendRequest
+     */
+    public void declineReceivedRequest(Long userId, FriendshipRequest friendRequest){
         Friendship friendship = findById(friendRequest.getId());
+        if(friendship.getUserId() != userId && friendship.getFriendId() != userId) {
+            throw new BadRequestException("요청에 속하지 않는 사용자입니다.");
+        }
         if(friendship.getStatus() == 1){
             throw new BadRequestException("요청을 이미 수락했습니다.");
         }
         friendRepository.delete(friendship);
     }
 
+    /**
+     * 받은 요청을 수락한다.
+     * @param friendRequest
+     */
     @Transactional
-    public void acceptReceivedRequest(Long id, FriendRequest friendRequest){
+    public void acceptReceivedRequest(Long userId, FriendshipRequest friendRequest){
         // 받은 요청을 수락한다.
         Friendship friendship = findById(friendRequest.getId());
+        if(friendship.getUserId() != userId && friendship.getFriendId() != userId) {
+            throw new BadRequestException("요청에 속하지 않는 사용자입니다.");
+        }
         if(friendship.getStatus() == 1){
             throw new BadRequestException("상대와 이미 친구입니다.");
         }
         friendRequest.updateFriend(friendship, 1);
     }
 
-    public void cancelSentRequest(Long id, FriendRequest friendRequest){
+    /**
+     * 보낸 요청을 취소한다.
+     * @param friendRequest
+     */
+    public void cancelSentRequest(Long userId, FriendshipRequest friendRequest){
         // 보낸 요청을 취소한다.
         Friendship friendship = findById(friendRequest.getId());
+        if(friendship.getUserId() != userId && friendship.getFriendId() != userId) {
+            throw new BadRequestException("요청에 속하지 않는 사용자입니다.");
+        }
         if(friendship.getStatus() == 1){
             throw new BadRequestException("상대가 요청을 이미 받았습니다.");
         }
